app1.1/app/services/report_service/requirements.txt
fastapi
uvicorn
aiofiles
requests
python-docx
python-multipart